+++ bootstrap/interface/__init__.py
@@ -36,7 +36,7 @@ from pydantic import (
 from typing_extensions import Annotated, NotRequired, Self, TypedDict
 
 from omc4py import TypeName, VariableName, exception, open_session
-from omc4py.v_1_21 import AsyncSession  # NOTE: update to latest
+from omc4py.v_1_22 import AsyncSession  # NOTE: update to latest
 
 from ..util import QueueingIteration, ensure_cancel, ensure_terminate