+++ omc4py/__init__.py
@@ -73,6 +73,7 @@ if TYPE_CHECKING:
         v_1_19,
         v_1_20,
         v_1_21,
+        v_1_22,
     )
 
     Command = Union[str, PathLike[str]]
@@ -81,7 +82,13 @@ if TYPE_CHECKING:
 def _select_session_type(
     version: Tuple[int, int]
 ) -> Tuple[type[BasicSession[Synchronous]], type[BasicSession[Asynchronous]]]:
-    if (1, 21) <= version:
+    if False:
+        pass
+    elif (1, 22) <= version:
+        from . import v_1_22
+
+        return v_1_22.Session, v_1_22.AsyncSession
+    elif (1, 21) <= version:
         from . import v_1_21
 
         return v_1_21.Session, v_1_21.AsyncSession
@@ -124,7 +131,28 @@ def _select_session_type(
 def open_session(
     omc_command: Command | None = None,
     *,
-    version: T[L[1], L[23, 22, 21]] | None = None,
+    version: T[L[1], L[22]] | T[L[1], L[23]] | T[L[1], L[24]] | None = None,
+    asyncio: Literal[False] = False,
+) -> v_1_22.Session:
+    ...
+
+
+@overload
+def open_session(
+    omc_command: Command | None = None,
+    *,
+    version: T[L[1], L[22]] | T[L[1], L[23]] | T[L[1], L[24]] | None = None,
+    asyncio: Literal[True],
+) -> v_1_22.AsyncSession:
+    ...
+
+
+# v1.21
+@overload
+def open_session(
+    omc_command: Command | None = None,
+    *,
+    version: T[L[1], L[21]],
     asyncio: Literal[False] = False,
 ) -> v_1_21.Session:
     ...
@@ -134,7 +162,7 @@ def open_session(
 def open_session(
     omc_command: Command | None = None,
     *,
-    version: T[L[1], L[23, 22, 21]] | None = None,
+    version: T[L[1], L[21]],
     asyncio: Literal[True],
 ) -> v_1_21.AsyncSession:
     ...
@@ -292,8 +320,7 @@ def open_session(
 def open_session(
     omc_command: Command | None = None,
     *,
-    version: T[L[1], L[13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 0]]
-    | T[L[0], int],
+    version: T[L[1], L[13]],
     asyncio: Literal[False] = False,
 ) -> v_1_13.Session:
     ...
@@ -303,8 +330,7 @@ def open_session(
 def open_session(
     omc_command: Command | None = None,
     *,
-    version: T[L[1], L[13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 0]]
-    | T[L[0], int],
+    version: T[L[1], L[13]],
     asyncio: Literal[True],
 ) -> v_1_13.AsyncSession:
     ...